refactor: #95 - buildSignupRequestPayload 이름 수정
